Add comprehensive form and UI components
This commit adds a comprehensive form to the 'solicita' page along with multiple UI components like `Checkbox`, `Tooltip`, `Select` etc. Changes also include addition of multiple icons and important details on the dashboard in the `TaxpayerPage`. The `.gitignore` file was updated to prevent tracking of JetBrains IDE settings. Moreover, the styling of the body in `layout.tsx` was simplified.