Add new terms and conditions page using MDX
Added a new terms and conditions page for a web application using MDX, introducing new MDX dependencies for rendering and a custom 'mdx-components.tsx' for styling. The added terms and conditions have been written in Spanish and are housed in 'page.mdx'. A new layout file was also added for the MDX pages. State of account card has been updated to display a different total amount due property.